+++ src/Microsoft.NetCore.Analyzers/Core/Security/xlf/SystemSecurityCryptographyResources.pl.xlf
@@ -166,6 +166,21 @@
         <target state="translated">Element {0} używa złamanych algorytmów kryptograficznych {1}</target>
         <note />
       </trans-unit>
+      <trans-unit id="DoNotUseDSA">
+        <source>Do Not Use Digital Signature Algorithm (DSA)</source>
+        <target state="new">Do Not Use Digital Signature Algorithm (DSA)</target>
+        <note />
+      </trans-unit>
+      <trans-unit id="DoNotUseDSADescription">
+        <source>DSA is an asymmetric algorithms that are weak. Switch to an RSA with at least 2048 key size, ECDH or ECDSA algorithm instead.</source>
+        <target state="new">DSA is an asymmetric algorithms that are weak. Switch to an RSA with at least 2048 key size, ECDH or ECDSA algorithm instead.</target>
+        <note />
+      </trans-unit>
+      <trans-unit id="DoNotUseDSAMessage">
+        <source>Weak asymmetric encryption algorithm {0} is banned by the SDL</source>
+        <target state="new">Weak asymmetric encryption algorithm {0} is banned by the SDL</target>
+        <note />
+      </trans-unit>
       <trans-unit id="DoNotUseDeprecatedSecurityProtocols">
         <source>Do Not Use Deprecated Security Protocols</source>
         <target state="translated">Nie używaj przestarzałych protokołów zabezpieczeń</target>
@@ -214,21 +229,6 @@
       <trans-unit id="DoNotUseSHA1Description">
         <source>This type implements SHA1, a cryptographically insecure hashing function. Hash collisions are computationally feasible for the SHA-1 and SHA-0 algorithms. Replace this usage with a SHA-2 family hash algorithm (SHA512, SHA384, SHA256).</source>
         <target state="translated">Ten typ implementuje niezabezpieczoną kryptograficznie funkcję skrótu SHA1. Istnieje możliwość wystąpienia kolizji skrótów w przypadku algorytmów SHA-1 i SHA-0. Zastąp to użycie algorytmem wartości skrótu z rodziny SHA-2 (SHA512, SHA384, SHA256).</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="DoNotUseWeakAsymmetricAlgorithm">
-        <source>Do Not Use Weak Asymmetric Algorithm</source>
-        <target state="new">Do Not Use Weak Asymmetric Algorithm</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="DoNotUseWeakAsymmetricAlgorithmDescription">
-        <source>Finds creation of asymmetric algorithms that are weak or have insufficient key size.</source>
-        <target state="new">Finds creation of asymmetric algorithms that are weak or have insufficient key size.</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="DoNotUseWeakAsymmetricAlgorithmMessage">
-        <source>Weak asymmetric encryption algorithm {0} is banned by the SDL. Switch to an RSA with at least 2048 key size, ECDH or ECDSA algorithm instead</source>
-        <target state="new">Weak asymmetric encryption algorithm {0} is banned by the SDL. Switch to an RSA with at least 2048 key size, ECDH or ECDSA algorithm instead</target>
         <note />
       </trans-unit>
       <trans-unit id="DoNotUseWeakCryptographicAlgorithms">
